Base iteration upload
Contains the base code including the SQL schema for a database, the ability to add new scripts and players into the database.

This version also can assign roles to each player in order to balance the teams evenly, based on character strength, player ELOs for the good and evil team, historical data, and how frequently the player has been on the good/evil team.